Fix complex interface (#74)
As it is the complex interface is not working, it gives the wrong answer when we use the field K.scomplex. 

The problem is an error in the construction of the QR matrix: to compute 2*real(m(i,j)) it does m(i,j) - m(j,i) instead of m(i,j) + m(j,i). As such this only affects problems where the real part of the SDP coefficients are nonzero (which might be useful for testing). This PR fixes it. I'm happy to send code to demonstrate the issue, but I don't see how to attach a file to a PR.

I also fixed a typo in an error message that made me waste a lot of time with debugging.